requirements.txt
bcrypt==3.1.3
Cerberus==1.1
cffi==1.10.0
coverage==4.4.1
Django==1.11.1
psycopg2==2.7.3.2
pycparser==2.17
PyJWT==1.5.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0
